+++ src/main/java/gs/mclo/api/Util.java
@@ -14,23 +14,42 @@ import java.util.zip.GZIPInputStream;
 
 public class Util {
 
+    /**
+     * List all files in a directory that match the allowed file name pattern
+     * @param directory directory to list files from
+     * @return array of file names that can be uploaded
+     */
     public static String[] listFilesInDirectory(File directory) {
-        if(!directory.exists())
+        if (!directory.exists()) {
             return new String[0];
+        }
+
         String[] files = directory.list();
         if (files == null)
             files = new String[0];
 
         return Arrays.stream(files)
-                .filter(f -> f.matches(Log.ALLOWED_FILE_NAME_PATTERN.pattern()))
+                .filter(f -> f.matches(LogReader.ALLOWED_FILE_NAME_PATTERN.pattern()))
                 .sorted()
                 .toArray(String[]::new);
     }
 
+    /**
+     * List all files in a directory that match the allowed file name pattern
+     * @param directory directory to list files from
+     * @return array of file names that can be uploaded
+     */
     public static String[] listFilesInDirectory(Path directory) {
         return listFilesInDirectory(directory.toFile());
     }
 
+    /**
+     * Parse the response body as JSON
+     * @param clazz class to parse the JSON into
+     * @param gson gson instance
+     * @return body handler
+     * @param <T> class type
+     */
     public static <T> HttpResponse.BodyHandler<T> parseResponse(Class<T> clazz, Gson gson) {
         return responseInfo -> HttpResponse.BodySubscribers.mapping(HttpResponse.BodySubscribers.ofString(StandardCharsets.UTF_8),
                 body -> gson.fromJson(body, clazz));